requirements.train.txt
ultralytics==8.3.10
datasets==2.20.0
wandb==0.17.7
google-cloud-storage==2.18.2
gcsfs==2024.6.1
numpy==1.26.4
pillow==10.4.0
pyyaml==6.0.2
opencv-python-headless==4.10.0.84
